requirements.txt
amqp==5.2.0
annotated-types==0.6.0
anyio==4.3.0
asgiref==3.8.1
billiard==4.2.0
black==24.10.0
boto3==1.34.60
botocore==1.34.134
celery==5.3.6
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
crispy-bootstrap5==2024.2
cryptography==42.0.5
defusedxml==0.8.0rc2
distro==1.9.0
dj-database-url==2.1.0
Django==5.1
django-crispy-forms==2.1
django-redis==5.4.0
django-storages==1.14.2
djangorestframework==3.15.2
djangorestframework-api-key==3.0.0
gunicorn==22.0.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
idna==3.7
jmespath==1.0.1
kombu==5.3.7
libgravatar==1.0.4
mypy-extensions==1.0.0
mysqlclient==2.2.4
numpy==1.26.4
oauthlib==3.2.2
openai==1.23.1
opencv-python==4.9.0.80
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.3.6
prompt-toolkit==3.0.43
psycopg2==2.9.9
pycparser==2.21
pydantic==2.6.3
pydantic_core==2.16.3
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
redis==5.0.3
requests==2.32.3
requests-oauthlib==1.4.0
s3transfer==0.10.0
six==1.16.0
sniffio==1.3.1
social-auth-app-django==5.4.1
social-auth-core==4.5.4
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.6.0
